Rendering fix
